Changed Array[Byte] to a type parameter.
This is mostly to make it easier to write tests, using the in-memory
sorted map. Having lots of trouble convincing the compiler that
Strings should be Ordered. Gave up and went with Comparable, and
changed comparisons to use ugly .compareTo instead. I don't understand
what happened to RichString in 2.9, and I don't understand how to use
StringOps. I have some studying to do.