Small redesign to use ValueRecords.
To make traverse easy, I'm now putting all values (flushed or not)
into a single map, where the value are ValueRecords. These are
essentially like an Either[Offset, ValueType], but mutable to make
it easier to deal with flushing multiple keys referring to the same
value in a single pass.