Starting to add a few "real" tests.
I'm refining the design as I go - I think I might merge the valueHashes
map into the ValueRecord type, because I realized that when I put an
existing key, I need to remove the key from the set associated with
the hash of its old value. Having that set in the record would make it
easier to get at.